add report month by course 1

+++ src/main/java/com/marcedev/attendance/dto/CourseMonthlyAttendanceDTO.java
@@ -5,10 +5,10 @@ public class CourseMonthlyAttendanceDTO {
     private Long studentId;
     private String studentName;
     private Long present;
-    private Double totalClasses; // <── CAMBIÓ A DOUBLE
+    private Long totalClasses; // <── CAMBIÓ A DOUBLE
     private Double percent;
 
-    public CourseMonthlyAttendanceDTO(Long studentId, String studentName, Long present, Double totalClasses, Double percent) {
+    public CourseMonthlyAttendanceDTO(Long studentId, String studentName, Long present, Long totalClasses, Double percent) {
         this.studentId = studentId;
         this.studentName = studentName;
         this.present = present;
@@ -19,6 +19,6 @@ public class CourseMonthlyAttendanceDTO {
     public Long getStudentId() { return studentId; }
     public String getStudentName() { return studentName; }
     public Long getPresent() { return present; }
-    public Double getTotalClasses() { return totalClasses; }
+    public Long getTotalClasses() { return totalClasses; }
     public Double getPercent() { return percent; }
 }
